fix: derive npc and character hitboxes from actor model
