+++ apps/game/server/npc/rpc.ts
@@ -3,8 +3,9 @@ import { ctxAreaLookup } from "../area/lookup";
 import { rpc } from "../rpc";
 import { roles } from "../user/auth";
 import { defineRoles } from "../user/define-roles";
+import { ctxActorModelLookup } from "../traits/appearance";
 import { ctxNpcService } from "./service";
-import { spawnNpcInstance } from "./npc-spawn-behavior";
+import { NpcSpawner } from "./npc-spawner";
 
 export const npcRoles = defineRoles("npc", ["spawnRandom"]);
 
@@ -15,13 +16,12 @@ export const npcRouter = rpc.router({
     .mutation(async ({ ctx }) => {
       const npcService = ctx.get(ctxNpcService);
       const state = ctx.get(ctxGameStateMachine);
-      const areas = ctx.get(ctxAreaLookup);
       const [{ npc, spawn }] = await npcService.getAllSpawnsAndTheirNpcs();
-      const area = areas.get(spawn.areaId);
-      if (!area) {
-        throw new Error(`Area not found: ${spawn.areaId}`);
-      }
-      const instance = spawnNpcInstance(npc, spawn, area);
+      const spawner = new NpcSpawner(
+        ctx.get(ctxAreaLookup),
+        ctx.get(ctxActorModelLookup),
+      );
+      const instance = spawner.spawnInstance(npc, spawn);
       state.actors.set(instance.id, { type: "npc", ...instance });
     }),
 });
